requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.7
python-dotenv>=1.0
openai>=1.51.0
chromadb>=0.5.5
sentence-transformers>=3.0
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
# Torch stack (CPU)
torch>=2.3; platform_system!="Darwin" or platform_machine!="arm64"
torchvision>=0.18; platform_system!="Darwin" or platform_machine!="arm64"
# macOS arm64 users: установите torch вручную по инструкции PyTorch
Pillow>=10.3
open-clip-torch>=2.24.0
python-multipart>=0.0.9
